Fix 2.9 hybrid compatibility
All broken because we moved gateway files to sub package
Move RCP structures to separate package, so they can be imported by MDCB
